Apply code review suggestions to test code.
